requirements.txt
# --- Scraping ---
bs4==0.0.2
selenium==4.18.1
webdriver-manager==4.0.1
undetected-chromedriver==3.5.5
setuptools>=70.0.0
lxml==5.1.0
urllib3==2.2.1
requests==2.31.0
tenacity==8.2.3

# --- Data ---
pandas==2.2.1
numpy==1.26.4

# --- Sentiment Analysis ---
nltk==3.8.1
textblob==0.18.0
vaderSentiment==3.3.2

# --- Price Forecasting ---
prophet==1.1.5
pystan==2.19.1.1
cmdstanpy==1.2.0

# --- Visualisation ---
plotly==5.20.0
streamlit==1.32.0

# --- Database ---
pymongo==4.6.2

# --- Alerts ---
python-dotenv==1.0.1

# --- Server ---
flask-cors==4.0.0
gunicorn==21.2.0

-e .